Extract App::Generate::Command
