+++ lib/hanami/cli/commands/app/generate/operation.rb
@@ -13,34 +13,16 @@ module Hanami
         module Generate
           # @since 2.2.0
           # @api private
-          class Operation < App::Command
-            argument :name, required: true, desc: "Operation name"
-            option :slice, required: false, desc: "Slice name"
-
+          class Operation < Generate::Command
             example [
               %(books.add               (MyApp::Books::Add)),
               %(books.add --slice=admin (Admin::Books::Add)),
             ]
-            attr_reader :generator
-            private :generator
 
             # @since 2.2.0
             # @api private
-            def initialize(
-              fs:, inflector:,
-              generator: Generators::App::Operation.new(fs: fs, inflector: inflector),
-              **opts
-            )
-              super(fs: fs, inflector: inflector, **opts)
-              @generator = generator
-            end
-
-            # @since 2.2.0
-            # @api private
-            def call(name:, slice: nil, **)
-              slice = inflector.underscore(Shellwords.shellescape(slice)) if slice
-
-              generator.call(app.namespace, name, slice)
+            def initialize(**opts)
+              super(generator_class: Generators::App::Operation, **opts)
             end
           end
         end
