+++ spec/unit/hanami/cli/commands/app/generate/operation_spec.rb
@@ -3,12 +3,12 @@
 require "hanami"
 
 RSpec.describe Hanami::CLI::Commands::App::Generate::Operation, :app do
-  subject { described_class.new(fs: fs, inflector: inflector, generator: generator) }
+  subject { described_class.new(fs: fs, inflector: inflector, generator_class: generator_class, out: out) }
 
   let(:out) { StringIO.new }
   let(:fs) { Hanami::CLI::Files.new(memory: true, out: out) }
   let(:inflector) { Dry::Inflector.new }
-  let(:generator) { Hanami::CLI::Generators::App::Operation.new(fs: fs, inflector: inflector, out: out) }
+  let(:generator_class) { Hanami::CLI::Generators::App::Operation }
   let(:app) { Hanami.app.namespace }
   let(:dir) { inflector.underscore(app) }
